+++ src/kernels/ft_to_vis.cl
@@ -44,7 +44,7 @@ __kernel void ft_to_vis(
     __private unsigned int n_vis,
     __global float * output)
 {
-    int i = get_global_id(0);
+    size_t i = get_global_id(0);
     // Lookup the index of the UV point which corresponds to this data point.
     unsigned int uv_index = uv_ref[i];
     // Get the Complex values.
